// client/src/app/components/elements.tsx
// Will be remove after the API is ready: (openApi)
export interface Tournament {
    id: number;
    name: string;
    sport: string;
    sex: string;
    ageGroup: string[];
    category: string;
    startDate: string;
    endDate: string;
    fees: number;
    location: string;
    description: string;
}